Create produce_data.py
When training a neural network, we need to present it with a set of problem in- stances - the graphs - and a set of solutions - the corresponding chromatic numbers. These need to be encoded as strings over an alphabet that the computer can “understand”.
Our data is a function of the graph’s number of vertices (size) and of the probability of an edge connecting to vertices (p). We produce our graphs us- ing the Sage RandomGNP generator function. We enforce canonical labelling to ensure that any two isomorphic graphs have the same labelling.
We “translate” our graphs into an alphabet that is readable by our network by considering the graphs’ adjacency matrix, that is, a size-dimensional square matrix with a 1 in entry (j, k) if vertices j and k are connected by an edge, and a 0 otherwise.
We immediately notice that, due to the definition of adjacency matrix, M will always be symmetric. Furthermore, as we do not allow self-loops, the diag- onal will always be filled with zeroes. This means that all the information is carried in the entries above the diagonal. We therefore get rid of all the redundant information and hopefully speed up our network’s computations by only considering the entries above and including the first upper diagonal. This is done with the aid of the numpy function triu_indices.
The function produce_data then returns a tuple called training_data, which includes two lists: the first contains 100.000 np.arrays which carry the information about the graphs’ adjacency matrices, formatted as above, the second contains the respective graph chromatic numbers. validation_data and test_data are produced in the exact same way, with the unique difference that we only produce 10.000 training examples for each tuple.